pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pactum"
version = "0.1.0"
description = "AI Contracts — first-class, versioned, testable interfaces and deterministic runtimes for AI components."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [
    { name = "Pactum Contributors" }
]
keywords = ["ai", "contracts", "testing", "deterministic", "llm", "runtime"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "click>=8.1",
    "pyyaml>=6.0",
    "openai>=1.0",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-asyncio>=0.21",
]

[project.scripts]
pactum = "pactum.cli.main:cli"

[project.urls]
Homepage = "https://github.com/pactum-ai/pactum"
Documentation = "https://github.com/pactum-ai/pactum/tree/main/docs"
Repository = "https://github.com/pactum-ai/pactum"
Issues = "https://github.com/pactum-ai/pactum/issues"

[tool.setuptools.packages.find]
include = ["pactum*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
